# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a simple API Expressauth",
    "version": "1.0.0",
    "title": "ExpressAuth",
    "termsOfService": "http://xprsau.io/terms/",
    "contact": {
      "email": "teamapi@xprsau.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3030",
  "basePath": "/api/",
  "tags": [
    {
      "name": "Users",
      "description": "User auth register & login",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://xprsau.io/api-user"
      }
    },
    {
      "name": "Products",
      "description": "Create and Get Product",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://xprsau.io/api-product"
      }
    }
  ],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Register new user",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to register",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "Registration success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/signin": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Login User",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to login",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User login success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/product": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Get all produts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Create new product",
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "Product that we want to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Product created success",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User":{
      "required":[
        "username",
        "password"
      ],
      "properties":{
        "username":{
          "type": "string"
        },
        "password":{
          "type": "string"
        }
      }
    },
    "Users":{
      "type":"array",
      "$ref": "#/definitions/User"
    },
    "Product":{
      "required":[
        "prod_name",
        "prod_desc",
        "prod_price"
      ],
      "properties":{
        "prod_name":{
          "type": "string"
        },
        "prod_desc":{
          "type": "string"
        },
        "prod_price":{
          "type": "double"
        }
      }
    },
    "Products":{
      "type":"array",
      "$ref": "#/definitions/Product"
    }
  }
}